+++ src/L.Control.HtmlLegend.js
@@ -18,6 +18,7 @@ L.Control.HtmlLegend = L.Control.extend({
         collapsedOnInit: false,
 
         disableVisibilityControls: false,
+        updateOpacity: null,
         defaultOpacity: 1,
         visibleIcon: 'leaflet-html-legend-icon-eye',
         hiddenIcon: 'leaflet-html-legend-icon-eye-slash',
@@ -27,6 +28,8 @@ L.Control.HtmlLegend = L.Control.extend({
     onAdd(map) {
         this._map = map;
         this._container = L.DomUtil.create('div', 'leaflet-control leaflet-bar leaflet-html-legend');
+        this._lastId = 0;
+        this._entries = {};
 
         // Disable events on container
         L.DomEvent.disableClickPropagation(this._container);
@@ -40,15 +43,25 @@ L.Control.HtmlLegend = L.Control.extend({
     render() {
         L.DomUtil.empty(this._container);
 
-        this.options.legends.forEach(this._renderLegend, this);
+        this.options.legends.forEach(legend => this._renderLegend(legend), this);
 
         this._checkVisibility();
     },
 
     addLegend(legend) {
-        this.options.legends.push(legend);
         if (this._map) {
             this._renderLegend(legend);
+        }
+    },
+
+    removeLegend(itemIdx) {
+        const entry = this._entries[itemIdx]
+        if (entry) {
+            if (entry.layer && entry.events) {
+                Object.entries(entry.events).forEach(([event, handler]) => entry.layer.off(event, handler))
+            }
+            L.DomUtil.remove(this._entries[itemIdx].container)
+            delete this._entries[itemIdx]
         }
     },
 
@@ -73,11 +86,13 @@ L.Control.HtmlLegend = L.Control.extend({
         }
 
         const block = L.DomUtil.create('div', className, this._container);
+        const entryIdx = ++this._lastId;
+        this._entries[entryIdx] = { container: block }
 
         if (this.options.collapseSimple && elements.length === 1 && !elements[0].label) {
             this._addElement(elements[0].html, legend.name, elements[0].style, block);
-            this._connectLayer(block, legend);
-            return;
+            this._connectLayer(block, legend, entryIdx);
+            return block;
         }
 
         if (legend.name) {
@@ -105,7 +120,8 @@ L.Control.HtmlLegend = L.Control.extend({
             this._addElement(element.html, element.label, element.style, elementContainer);
         }, this);
 
-        this._connectLayer(block, legend);
+        this._connectLayer(block, legend, entryIdx);
+        return block;
     },
 
     _addElement(html, label, style, container) {
@@ -121,15 +137,28 @@ L.Control.HtmlLegend = L.Control.extend({
     },
 
     _updateOpacity(layer, opacity) {
-        if (typeof layer.setOpacity === 'function') {
+        if (typeof this.options.updateOpacity === 'function') {
+            this.options.updateOpacity(layer, opacity)
+        } else if (typeof layer.setOpacity === 'function') {
             layer.setOpacity(opacity);
-        }
-        else if (typeof layer.setStyle === 'function') {
+        } else if (typeof layer.setStyle === 'function') {
             layer.setStyle({ opacity });
         }
     },
 
-    _connectLayer(container, legend) {
+    _layerAdd(container) {
+        this._activeLayers += 1;
+        container.style.display = '';
+        this._checkVisibility();
+    },
+
+    _layerRemove(container) {
+        this._activeLayers -= 1;
+        container.style.display = 'none';
+        this._checkVisibility();
+    },
+
+    _connectLayer(container, legend, entryIdx) {
         const layer = legend.layer;
 
         if (!layer) {
@@ -186,19 +215,15 @@ L.Control.HtmlLegend = L.Control.extend({
             L.DomUtil.create('i', this.options.hiddenIcon, opacityController);
         };
 
-        this._map.on('layeradd', (e) => {
-            if (e.layer === layer) {
-                this._activeLayers += 1;
-                container.style.display = '';
-                this._checkVisibility();
-            }
-        }).on('layerremove', (e) => {
-            if (e.layer === layer) {
-                this._activeLayers -= 1;
-                container.style.display = 'none';
-                this._checkVisibility();
-            }
-        });
+
+        const layerAdd = this._layerAdd.bind(this, container)
+        const layerRemove = this._layerRemove.bind(this, container)
+        layer.on('add', layerAdd).on('remove', layerRemove)
+        this._entries[entryIdx].layer = layer
+        this._entries[entryIdx].events = {
+            add: layerAdd,
+            remove: layerRemove
+        }
     },
 
     _checkVisibility() {